Handle MQTT connection drops
If the MQTT connection drops (e.g., the Loop server shuts down, or there is a network split), the system now behaves more predictably.

First, the top-level supervisor is now `rest_for_one`, meaning that as soon as the Announcer loses its connection and terminates, all bundles will also terminate, only to come back once the Announcer has a connection (if the Announcer can't get a connection, there's no point having all the commands in all the bundles trying to get a connection as well).

Second, the Announcer no longer tries to manage reconnection internally, preferring instead to offload this to the OTP supervisor. This simplifies the Announcer and makes the behavior of the entire supervision tree more predictable. It also paves the way to use OS-level supervision to restart Relay in the face of longer network splits.