Start up a message bus under test
Mix has a questionable habit of starting up the application before
running tests, even if you don't actually need or want the application
to be up. With these changes to how the Announcer starts, Relay as a
whole will not come up if there is no message bus for it to connect to,
and as a result, tests will not even run.

To get around this, we bring in emqttd as a test-only dependency, and
start it up as a dependency only if we're running tests. Test
configuration for the message bus was copied over from Cog.