Rework Announcer as an FSM
To be more robust in the face of network issues, the Announcer has
been converted from a server to an FSM. The overall scheme of the
state machine is:

    introduction -> snapshot announcement -> ongoing incremental announcements

Announcements happen strictly in that order after startup, in order to
ensure the bot has a consistent picture of what the Relay offers.

Announcements are also periodically re-sent until the bot sends back
an acknowledgment message. This ensures that no messages get lost,
further ensuring consistent state everywhere.

Additionally, existing bundle startup has been moved out of the
Announcer and into a Starter server. Admittedly, this is perhaps
overkill, but it ensures Single Responsibility and places startup into
the overall supervision tree in the appropriate place.