Rewrite of bundle installation
This commit dramatically reworks Relay's bundle installation
process. The rework was undertaken for the following reasons:

- Previous bundle installation was racy and, in certain circumstances,
  could wedge both Relay and Cog pretty tightly.
- Scanning for new bundles and installing them happened in a single
  process. Sharing a single process for all installs eliminates any
  fault isolation. One bad bundle could block all future
  installs, for example.

The new code jettisons the janky `Relay.Bundle.Installer` behaviour and
implementations of the same and replaces it with a master/worker process
arrangement between `Relay.Bundle.Scanner` and
`Relay.Bundle.Installer`. Installers run in separate processes and are
monitored by Scanner. If one crashes Scanner can pick up the pieces and
clean up the failed install. This also permits Relay to install multiple
bundles in parallel.